dev tanmay

+++ database/migrations/2025_04_04_080045_add_item_id_to_comments_table.php
@@ -11,16 +11,27 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('comments', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->foreignId('parent_id')->nullable()->constrained('comments')->onDelete('cascade');
-            $table->text('body');
-            $table->string('item_id')->nullable();
-            $table->morphs('commentable');
-            $table->softDeletes();
-            $table->timestamps();
+        Schema::table('comments', function (Blueprint $table) {
+            if (!Schema::hasColumn('comments', 'user_id')) {
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+            }
+            if (!Schema::hasColumn('comments', 'parent_id')) {
+                $table->foreignId('parent_id')->nullable()->constrained('comments')->onDelete('cascade');
+            }
+            if (!Schema::hasColumn('comments', 'body')) {
+                $table->text('body');
+            }
+            if (!Schema::hasColumn('comments', 'item_id')) {
+                $table->string('item_id')->nullable();
+            }
+            if (!Schema::hasColumn('comments', 'commentable_type')) {
+                $table->morphs('commentable');
+            }
+            if (!Schema::hasColumn('comments', 'deleted_at')) {
+                $table->softDeletes();
+            }
         });
+
     }
 
     /**
@@ -29,7 +40,7 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('comments', function (Blueprint $table) {
-            //
+            $table->dropColumn(['user_id', 'parent_id', 'body', 'item_id', 'commentable_type', 'commentable_id', 'deleted_at']);
         });
     }
 };
